+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useCaseBlocksLogic/utils/useTabsToBlocksMap.tsx
@@ -100,6 +100,7 @@ export const useTabsToBlocksMap = ({
     kybRegistryInfoBlock,
     companySanctionsBlock,
     individualsUserProvidedBlock,
+    individualsRegistryProvidedBlock,
     ubosRegistryProvidedBlock,
     storeInfoBlock,
     websiteBasicRequirementBlock,
@@ -314,6 +315,9 @@ export const useTabsToBlocksMap = ({
       } satisfies Parameters<typeof createKycBlocks>[0][number];
     },
     [
+      getStatusFromCheckStatus,
+      mutateInitiateIndividualVerificationAndSendEmail,
+      onEditCollectionFlow,
       getStatusFromTags,
       getInitiateKycEvent,
       getInitiateSanctionsScreeningEvent,
@@ -406,6 +410,9 @@ export const useTabsToBlocksMap = ({
       } satisfies Parameters<typeof createKycBlocks>[0][number];
     },
     [
+      mutateInitiateIndividualVerificationAndSendEmail,
+      onEditCollectionFlow,
+      workflow?.workflowDefinition?.config?.language,
       workflow?.id,
       caseState.actionButtonsEnabled,
       workflow?.workflowDefinition?.config?.isInitiateKycEnabled,
@@ -447,12 +454,114 @@ export const useTabsToBlocksMap = ({
             documents: [...(director?.documents ?? []), ...(workflow?.context?.kycDocuments ?? [])],
           });
         }) ?? [],
-    [workflow, endUsers, directorToIndividualAdapter],
+    [workflow, endUsers, directorToIndividualAdapter, getStatusFromCheckStatus],
+  );
+
+  const personOfInterestToIndividualAdapter = useCallback(
+    ({
+      ballerineEntityId,
+      role,
+    }: NonNullable<
+      TWorkflowById['context']['entity']['data']['additionalInfo']['peopleOfInterest']
+    >[number]) => {
+      const {
+        id: _id,
+        amlHits,
+        individualVerificationsChecks,
+        ...personOfInterestEndUser
+      } = endUsers?.find(endUser => endUser.id === ballerineEntityId) ?? {};
+      const status = getStatusFromCheckStatus(individualVerificationsChecks?.status);
+      const kycSession = omitPropsFromObject(
+        individualVerificationsChecks?.data ?? {},
+        'invokedAt',
+        'error',
+        'name',
+        'status',
+        'isRequestTimedOut',
+      );
+
+      return {
+        status,
+        documents: [],
+        kycSession,
+        aml: {
+          vendor: amlHits?.find(aml => !!aml.vendor)?.vendor,
+          hits: amlHits,
+        },
+        entityData: {
+          ...personOfInterestEndUser,
+          role,
+        },
+        isActionsDisabled: true,
+        isLoadingReuploadNeeded: false,
+        isLoadingApprove: false,
+        onInitiateKyc: () => {
+          if (!workflow?.id) {
+            console.error('No workflow id found');
+            toast.error('Something went wrong. Please try again later.');
+
+            return;
+          }
+
+          return mutateInitiateIndividualVerificationAndSendEmail({
+            endUserId: ballerineEntityId,
+            ongoingMonitoring: false,
+            withAml: true,
+            workflowRuntimeDataId: workflow?.id,
+            vendor: 'veriff',
+            language: workflow?.workflowDefinition?.config?.language ?? 'en',
+          });
+        },
+        onInitiateSanctionsScreening: () => {},
+        onApprove:
+          ({ ids }: { ids: string[] }) =>
+          () => {},
+        onReuploadNeeded:
+          ({ reason, ids }: { reason: string; ids: string[] }) =>
+          () => {},
+        onEdit: onEditCollectionFlow({ steps: ['company_ownership'] }),
+        reasons: [],
+        isReuploadNeededDisabled: true,
+        isApproveDisabled: true,
+        isInitiateKycDisabled: [
+          !workflow?.id,
+          !caseState.actionButtonsEnabled,
+          !workflow?.workflowDefinition?.config?.isInitiateKycEnabled,
+        ].some(Boolean),
+        isInitiateSanctionsScreeningDisabled: true,
+        isEditDisabled: [
+          !caseState.actionButtonsEnabled,
+          !workflow?.workflowDefinition?.config?.isKycEndUserEditEnabled,
+        ].some(Boolean),
+      } satisfies Parameters<typeof createKycBlocks>[0][number];
+    },
+    [
+      workflow?.id,
+      caseState.actionButtonsEnabled,
+      workflow?.workflowDefinition?.config?.isInitiateKycEnabled,
+      workflow?.workflowDefinition?.config?.isKycEndUserEditEnabled,
+      endUsers,
+      getStatusFromCheckStatus,
+      mutateInitiateIndividualVerificationAndSendEmail,
+      onEditCollectionFlow,
+      workflow?.workflowDefinition?.config?.language,
+    ],
+  );
+
+  const peopleOfInterest = useMemo(
+    () =>
+      workflow?.context?.entity?.data?.additionalInfo?.peopleOfInterest?.map(
+        personOfInterestToIndividualAdapter,
+      ) ?? [],
+    [
+      workflow?.context?.entity?.data?.additionalInfo?.peopleOfInterest,
+      personOfInterestToIndividualAdapter,
+    ],
   );
 
   const individuals = useMemo(
-    () => [...childWorkflows, ...deDupedDirectors],
-    [childWorkflows, deDupedDirectors],
+    () => [...childWorkflows, ...deDupedDirectors, ...peopleOfInterest],
+    [childWorkflows, deDupedDirectors, peopleOfInterest],
   );
 
   const kycBlocks = useKYCBlocks(individuals);
@@ -490,6 +599,7 @@ export const useTabsToBlocksMap = ({
     [Tab.DOCUMENTS]: [...businessDocumentBlocks],
     [Tab.INDIVIDUALS]: [
       ...individualsUserProvidedBlock,
+      ...individualsRegistryProvidedBlock,
       ...amlWithContainerBlock,
       ...mainRepresentativeBlock,
       ...uboDocumentBlocks,