+++ services/workflows-service/src/workflow/workflow.service.unit.test.ts
@@ -68,6 +68,12 @@ class FakeAssessmentsService extends BaseFakeRepository {
   }
 }
 
+class FakeKycService extends BaseFakeRepository {
+  constructor() {
+    super(Object);
+  }
+}
+
 const buildWorkflowDeifintion = (sequenceNum: number) => {
   return {
     id: sequenceNum.toString(),
@@ -128,6 +134,7 @@ describe('WorkflowService', () => {
   let testingModule: TestingModule;
   let workflowLogService;
   let assessmentsService;
+  let kycService;
 
   const configService = {
     WEBHOOK_URL: 'https://example.com',
@@ -162,6 +169,7 @@ describe('WorkflowService', () => {
     uiDefinitionService = new FakeUiDefinitionService();
     workflowLogService = new FakeWorkflowLogService();
     assessmentsService = new FakeAssessmentsService();
+    kycService = new FakeKycService();
     fakeHttpService = {
       requests: [],
 
@@ -226,6 +234,7 @@ describe('WorkflowService', () => {
       {} as any,
       workflowLogService,
       assessmentsService,
+      kycService,
     );
   });